got hh ll l h to work with d and i. now working on changing unsigned char

+++ libft/libft.h
@@ -103,7 +103,7 @@ void	ft_lstprint(t_list **alst);
 void	ft_lstprint_size(t_list **alst);
 char	**ft_free2d(char **dest);
 int		get_next_line(const int fd, char **line);
-char	*ft_itoa_base(int nbr, int len, int base);
+char	*ft_itoa_base(long long nbr, int len, int base);
 char	*ft_strrev(char *str);
 
 #endif
